+++ locus-api-android/src/main/java/locus/api/android/utils/LocusUtils.kt
@@ -152,9 +152,8 @@ object LocusUtils {
                     versions.add(lv)
                 }
             } catch (ignored: PackageManager.NameNotFoundException) {
-                Logger.logD(TAG, "getAvailableVersions($ctx), " +
-                        "e: $ignored")
-                //ignored.printStackTrace()
+//                Logger.logD(TAG, "getAvailableVersions($ctx), " +
+//                        "e: $ignored")
             }
         }